+++ CHANGELOG.md
@@ -5,6 +5,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.2] - 2020-07-22
+
+### Added
+- Added titles
+- Radar button to include cutoff lines in plot
+- Figure
+- Button to download the plot
+- Global boolean to check whether files were imported successfully
+- Added import file validation steps
+
+### Changed
+- Changed the title in all scripts
+- Updated README file
+- Plot layout
+
+### Removed
+- Removed the trend tab that showed the trending of all probes
+
 ## [0.0.1] - 2020-07-22
 
 ### Added